requirements.txt
pyyaml>=6.0
kubernetes>=24.2.0
boto3>=1.24.0
azure-identity>=1.11.0
azure-mgmt-containerservice>=20.0.0
click>=8.1.3
rich>=12.5.1
pytest>=7.0.0
pytest-mock>=3.8.0
